+++ CHANGELOG.md
@@ -9,6 +9,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## v6.3.0
+
+### Added
+
+- Add constants `CoordinateSystem*Well::POSITIONS_COUNT`
+- Add constants `MicroplateSet*::PLATE_COUNT`
+
 ## v6.2.0
 
 ### Added